% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etl_appsilon_requirement.R
\name{etl_appsilon_requirement}
\alias{etl_appsilon_requirement}
\title{ETL for Appsilon requirement}
\usage{
etl_appsilon_requirement(ships_data)
}
\arguments{
\item{ships_data}{Data.frame with columns (SHIPNAME, DATETIME, LON,LAT)}
}
\value{
data.frame
}
\description{
Objective:
Return one observation by each ship filtering by:
- The longest distance between two consecutive observations.(In meters).
}
\details{
and including:
- The previous and current coord.
- The time between the both observations (in seconds).
}
