% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etl_cleanse_ships_data.R
\name{etl_cleanse_ships_data}
\alias{etl_cleanse_ships_data}
\title{Cleanse shipping data}
\usage{
etl_cleanse_ships_data(df_raw_data)
}
\arguments{
\item{df_raw_data}{raw data provided by Appsilon}
}
\value{
data.frame
}
\description{
Applies necessary cleansing of the ships_data provided by Appsilon.
For internal use to update the pre-calculated data in the package.
For current requirements see details.
}
\details{
\itemize{
\item Removing duplicates
\item Removing instances of [SAT-AIS] and "SAT AIS"
\item Adding FLAG to the names of twenty-six ships with similar names, but different flags.
\item Correcting the presumed misspellings of certain ship-names.
\item Adding FLAG designation to "ARGO" and "C".
\item Arbitrarily adding a I & II designation to about ten ships with similar names, same flag, but different types and/or other characteristics.
\item Arbitrarily treating GAZPROM WEST and VOVAN as separate vessels (although we have our doubts).
}
}
