% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_dropdown.R
\name{mod_dropdown_server}
\alias{mod_dropdown_server}
\title{Dropdown Server Function}
\usage{
mod_dropdown_server(input, output, session, ships)
}
\arguments{
\item{input}{Shiny Input}

\item{output}{Shiny Output}

\item{session}{Shiny session}

\item{ships}{\itemize{
\item Data.frame with data to show in selects inputs.
}}
}
\description{
Dropdown Server Function
}
\author{
Pablo Pagnone
}
