% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_marine.R
\name{mod_marine_server}
\alias{mod_marine_server}
\title{marine Server Function}
\usage{
mod_marine_server(input, output, session, ships)
}
\arguments{
\item{ships}{- Data.frame with ships information}

\item{id, input, output, session}{Internal parameters for {shiny}.}
}
\description{
marine Server Function
}
\author{
Pablo Pagnone
}
