% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_marine.R
\name{mod_marine_ui}
\alias{mod_marine_ui}
\title{marine UI Function}
\usage{
mod_marine_ui(id)
}
\arguments{
\item{id, input, output, session}{Internal parameters for {shiny}.}
}
\description{
A shiny Module.
}
\author{
Pablo Pagnone
}
