% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ship_position_map.R
\name{ship_position_map}
\alias{ship_position_map}
\title{Map to show the ships position}
\usage{
ship_position_map(data, show_previous_position = FALSE)
}
\arguments{
\item{data}{Ships data.frame}
}
\value{
leaflet plot
}
\description{
Map to show the ships position
}
