% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_marine.R
\name{mod_marine_server}
\alias{mod_marine_server}
\title{Marine Server Function}
\usage{
mod_marine_server(input, output, session, ships)
}
\arguments{
\item{input}{Internal parameters for {shiny}.}

\item{output}{Internal parameters for {shiny}.}

\item{session}{Internal parameters for {shiny}.}

\item{ships}{\itemize{
\item Data.frame with ships information
}}
}
\description{
Marine Server Function
}
\author{
Pablo Pagnone
}
