% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_marine.R
\name{mod_marine_ui}
\alias{mod_marine_ui}
\title{Marine UI Function}
\usage{
mod_marine_ui(id)
}
\arguments{
\item{id}{Internal parameters for {shiny}.}
}
\description{
Marine UI Module, show dropdowns to select a vessel and show in
a map.
}
\author{
Pablo Pagnone
}
